more efficient f computing + bug fix

+++ woodelf/lts_polynomial_multiplication.py
@@ -10,7 +10,7 @@ def bits_matrix(x: np.ndarray, k: int) -> np.ndarray:
     returns: shape (k, n), rows are bits (k-1),...,1,0 (2^(k-1) down to LSB)
     """
     # ensure x is unsigned (np.uint) for fast bit ops
-    shifts = np.arange(k-1, -1, -1, dtype=np.uint64)[:, None]  # (5,1): 4,3,2,1,0
+    shifts = np.arange(k-1, -1, -1, dtype=np.uint8)[:, None]  # (5,1): 4,3,2,1,0
     return ((x[None, :] >> shifts) & 1).astype(np.uint8)
 
 def neg_bits_matrix(x: np.ndarray, k: int) -> np.ndarray:
@@ -18,7 +18,7 @@ def neg_bits_matrix(x: np.ndarray, k: int) -> np.ndarray:
     identical to bits_matrix(x,k).replace({0:1, 1:0})
     """
     # ensure x is unsigned (np.uint) for fast bit ops
-    shifts = np.arange(k-1, -1, -1, dtype=np.uint64)[:, None]  # (5,1): 4,3,2,1,0
+    shifts = np.arange(k-1, -1, -1, dtype=np.uint8)[:, None]  # (5,1): 4,3,2,1,0
     return (((x[None, :] >> shifts) + 1) & 1).astype(np.uint8)
